refactor(data): share one PostgREST pagination helper across read layers (#336)
The cardio read layer already had a `fetchAllRows` that pages around
PostgREST's 1000-row response cap; the weight-room fix in the previous
commit duplicated it. Move the implementation to `lib/data/paged-read.ts`
and have both assemblers use it, so the page size and the short-page
termination rule live in one place.

`cardio-shared` re-exports it, keeping `cardio-server`'s import path
unchanged. No behavior change — full suite still green.
